+++ src/Ext/Techno/Body.h
@@ -51,8 +51,9 @@ public:
 		// Used for Passengers.SyncOwner.RevertOnExit instead of TechnoClass::InitialOwner / OriginallyOwnedByHouse,
 		// as neither is guaranteed to point to the house the TechnoClass had prior to entering transport and cannot be safely overridden.
 		HouseClass* OriginalPassengerOwner;
-		bool HasCarryoverWarpInDelay;          // Converted from object with Teleport Locomotor to one with a different Locomotor while still phasing in.
+		bool HasRemainingWarpInDelay;          // Converted from object with Teleport Locomotor to one with a different Locomotor while still phasing in OR set if ChronoSphereDelay > 0.
 		int LastWarpInDelay;                   // Last-warp in delay for this unit, used by HasCarryoverWarpInDelay.
+		bool IsBeingChronoSphered;             // Set to true on units currently being ChronoSphered, does not apply to Ares-ChronoSphere'd buildings or Chrono reinforcements.
 
 		ExtData(TechnoClass* OwnerObject) : Extension<TechnoClass>(OwnerObject)
 			, TypeExtData { nullptr }
@@ -80,8 +81,9 @@ public:
 			, CanCurrentlyDeployIntoBuilding { false }
 			, FiringObstacleCell {}
 			, OriginalPassengerOwner {}
-			, HasCarryoverWarpInDelay { false }
+			, HasRemainingWarpInDelay { false }
 			, LastWarpInDelay { 0 }
+			, IsBeingChronoSphered { false}
 		{ }
 
 		void OnEarlyUpdate();